+++ enspara/cluster/util.py
@@ -17,13 +17,22 @@ from enspara.util.parallel import auto_nprocs
 from enspara import mpi
 import itertools
 import pickle
+import time
+from joblib import Parallel, delayed
+import psutil
+from functools import partial
+import resource
+
 
 import mdtraj as md
 import numpy as np
 
 from ..geometry.libdist import euclidean, manhattan
 
-from ..exception import ImproperlyConfigured, DataInvalid
+from enspara.util.load import (concatenate_trjs, sound_trajectory,
+                               load_as_concatenated)
+
+from  enspara.exception import ImproperlyConfigured, DataInvalid
 from ..ra.ra import partition_list, partition_indices
 from enspara import ra
 
@@ -346,7 +355,7 @@ def load_trajectories(topologies, trajectories, selections, stride, processes):
             # noop, but causes fast-fail w/bad args.atoms
             sentinel_trj.top.select(selection)
         except:
-            raise exception.ImproperlyConfigured((
+            raise ImproperlyConfigured((
                 "The provided selection '{s}' didn't match the topology "
                 "file, {t}").format(s=selection, t=top))
 
@@ -361,7 +370,7 @@ def load_trajectories(topologies, trajectories, selections, stride, processes):
 
         if n_inds is not None:
             if n_inds != len(indices):
-                raise exception.ImproperlyConfigured(
+                raise ImproperlyConfigured(
                     ("Selection on topology %s selected %s atoms, but "
                      "other selections selected %s atoms.") %
                     (topfile, len(indices), n_inds))
@@ -410,7 +419,7 @@ def load_asymm_frames(center_indices, trajectories, topology, subsample):
                 target_centers,
                 top=md.load(topfile).top,
                 stride=subsample)
-        except exception.ImproperlyConfigured:
+        except ImproperlyConfigured:
             logger.error('Failure to load cluster centers %s for topology %s',
                          topfile, target_centers)
             raise
@@ -538,3 +547,188 @@ def write_assignments_and_distances_with_reassign(result, args, intermediate_n=N
         ra.save(args.assignments, assig)
     else:
         logger.debug("Got --no-reassign, not doing reassigment")
+
+def compute_batches(lengths, batch_size):
+    """Compute batches (slices into lengths) of combined length at most
+    size batch_size.
+    """
+
+    batch_sizes = [[]]
+    batch_indices = [[]]
+    for i, l in enumerate(lengths):
+        if sum(batch_sizes[-1]) + l < batch_size:
+            batch_sizes[-1].append(l)
+            batch_indices[-1].append(i)
+        else:
+            batch_sizes.append([l])
+            batch_indices.append([i])
+
+    return batch_indices
+
+
+def determine_batch_size(n_atoms, dtype_bytes, frac_mem):
+    floats_per_frame = n_atoms * 3
+    bytes_per_frame = floats_per_frame * dtype_bytes
+
+    mem = psutil.virtual_memory()
+    bytes_total = mem.total
+
+    batch_size = int(bytes_total * frac_mem / bytes_per_frame)
+    batch_gb = batch_size * bytes_per_frame / (1024**3)
+
+    return batch_size, batch_gb
+
+
+def batch_reassign(targets, centers, lengths, frac_mem, n_procs=None):
+
+    example_center = centers[0]
+
+    DTYPE_BYTES = 4
+    batch_size, batch_gb = determine_batch_size(
+        example_center.n_atoms, DTYPE_BYTES, frac_mem)
+
+    logger.info(
+        'Batch max size set to %s frames (~%.2f GB, %.1f%% of total RAM).' %
+        (batch_size, batch_gb, frac_mem*100))
+
+    if batch_size < max(lengths):
+        raise ImproperlyConfigured(
+            'Batch size of %s was smaller than largest file (size %s).' %
+            (batch_size, max(lengths)))
+
+    batches = compute_batches(lengths, batch_size)
+
+    assignments = []
+    distances = []
+
+    for i, batch_indices in enumerate(batches):
+        tick = time.perf_counter()
+        logger.info("Starting batch %s of %s", i+1, len(batches))
+        batch_targets = [targets[i] for i in batch_indices]
+
+        with timed("Loaded frames for batch in %.1f seconds", logger.info):
+            batch_lengths, xyz = load_as_concatenated(
+                [tfile for tfile, top, aids in batch_targets],
+                lengths=[lengths[i] for i in batch_indices],
+                args=[{'top': top, 'atom_indices': aids}
+                      for t, top, aids in batch_targets],
+                processes=n_procs)
+
+        # mdtraj loads as float32, and load_as_concatenated should thus
+        # also load as float32. This should _never_ be hit, but there might be
+        # some platform-specific situation where double != float64?
+        assert xyz.dtype.itemsize == DTYPE_BYTES
+
+        trj = md.Trajectory(xyz, topology=example_center.top)
+
+        with timed("Precentered trajectories in %.1f seconds", logger.debug):
+            trj.center_coordinates()
+
+        with timed("Assigned trajectories in %.1f seconds", logger.debug):
+            batch_assignments, batch_distances = assign_to_nearest_center(
+                    trj, centers, partial(md.rmsd, precentered=True))
+
+        # clear memory of xyz and trj to allow cleanup to deallocate
+        # these large arrays; may help with memory high-water mark
+        with timed("Cleared array from memory in %.1f seconds", logger.debug):
+            xyz_size = xyz.size
+            del trj, xyz
+
+        assignments.extend(partition_list(batch_assignments, batch_lengths))
+        distances.extend(partition_list(batch_distances, batch_lengths))
+
+        logger.info(
+            "Finished batch %s of %s in %.1f seconds. Coordinates array had "
+            "memory footprint of %.2f GB (of memory high-water mark %.2f/%.2f "
+            "GB).",
+            i, len(batches), time.perf_counter() - tick,
+            xyz_size * DTYPE_BYTES / 1024**3,
+            resource.getrusage(resource.RUSAGE_SELF).ru_maxrss / 1024**2,
+            psutil.virtual_memory().total / 1024**3)
+
+    return assignments, distances
+
+
+def reassign(topologies, trajectories, atoms, centers, frac_mem=0.5):
+    """Reassign a set of trajectories based on a subset of atoms and centers.
+
+    Parameters
+    ----------
+    topologies : list
+        List of topologies corresponding to the trajectories to be
+        reassigned.
+    trajectories : list of lists
+        List of lists of tajectories to be loaded in batches and
+        reassigned.
+    atoms : list
+        List of MDTraj atom query strings. Each string is applied to the
+        corresponding topology to choose which atoms will be used for
+        the reassignment.
+    centers : md.Trajectory or list of trajectories
+        The atoms representing the centers to reassign to.
+    frac_mem : float, default=0.5
+        The fraction of main RAM to use for trajectories. A lower number
+        will mean more batches.
+    """
+
+    n_procs = auto_nprocs()
+
+    # check input validity
+    if len(topologies) != len(trajectories):
+        raise ImproperlyConfigured(
+            "Number of topologies (%s) didn't match number of sets of "
+            "trajectories (%s)." % (len(topologies), len(trajectories)))
+    if len(topologies) != len(atoms):
+        raise ImproperlyConfigured(
+            "Number of topologies (%s) didn't match number of atom selection "
+            "strings (%s)." % (len(topologies), len(atoms)))
+
+    # iteration across md.Trajectory is insanely slow. Do it only once here.
+    if isinstance(centers, md.Trajectory):
+        tick = time.perf_counter()
+        logger.info('Centers are an md.Trajectory. Creating trj-list to '
+                    'avoid repeated iteration.')
+        # using in-place copies to reduce memory usage (and for speed)
+        centers = [centers.slice(i, copy=False) for i in range(len(centers))]
+        logger.info('Built trj list in %.1f seconds.',
+                    time.perf_counter() - tick)
+
+    # precenter centers (there will be many RMSD calcs here)
+    for c in centers:
+        c.center_coordinates()
+
+    with timed("Reassignment took %.1f seconds.", logger.info):
+        # build flat list of targets
+        targets = []
+        for topfile, trjfiles, atoms in zip(topologies, trajectories, atoms):
+            t = md.load(topfile).top
+            atom_ids = t.select(atoms)
+            for trjfile in trjfiles:
+                assert os.path.exists(trjfile)
+                targets.append((trjfile, t, atom_ids))
+
+        # determine trajectory length
+        tick_sounding = time.perf_counter()
+        logger.info("Sounding dataset of %s trajectories and %s topologies.",
+                    sum(len(t) for t in trajectories), len(topologies))
+
+        lengths = Parallel(n_jobs=n_procs)(
+            delayed(sound_trajectory)(f) for f, _, _ in targets)
+
+        logger.info("Sounded %s trajectories with %s frames (median length "
+                    "%i frames) in %.1f seconds.",
+                    len(lengths), sum(lengths), np.median(lengths),
+                    time.perf_counter() - tick_sounding)
+
+        assignments, distances = batch_reassign(
+            targets, centers, lengths, frac_mem=frac_mem, n_procs=n_procs)
+
+    if all([len(assignments[0]) == len(a) for a in assignments]):
+        logger.info("Trajectory lengths are homogenous. Output will "
+                    "be np.ndarrays.")
+        assert all([len(distances[0]) == len(d) for d in distances])
+        return np.array(assignments), np.array(distances)
+    else:
+        logger.info("Trajectory lengths are heterogenous. Output will "
+                    "be ra.RaggedArrays.")
+        return ra.RaggedArray(assignments), ra.RaggedArray(distances)
